Refactors additional resource loading
Moves the logic for loading additional battle scene resources into a dedicated function.

This improves code organization and maintainability by isolating the resource loading process.
