+++ src/js/resource/update-battle-scene-resources/index.ts
@@ -1,18 +1,12 @@
 import { Player } from "gbraver-burst-core";
 
-import { CANVAS_IMAGE_CONFIGS } from "../canvas-image/configs";
-import { CUBE_TEXTURE_CONFIGS } from "../cube-texture/configs";
-import { GLTF_CONFIGS } from "../gltf/configs";
 import { Resources } from "../index";
-import { loadResources } from "../loading/load-resources";
 import { mergeResources } from "../loading/merge-resources";
-import { ResourceType } from "../resource-type";
-import { SOUND_CONFIGS } from "../sound/configs";
-import { TEXTURE_CONFIGS } from "../texture/configs";
 import { getAdditionalArmdozerIds } from "./get-additional-armdozer-ids";
 import { getPlayerArmdozerIds } from "./get-player-armdozer-ids";
 import { getResourceArmdozerIds } from "./get-resource-armdozer-ids";
 import { getResourceTypes } from "./get-resource-types";
+import { loadAdditionalResources } from "./load-additional-resources";
 
 /**
  * バトルシーンのためにリソースの追加読み込み、破棄を行う
@@ -33,18 +27,9 @@ export async function updateBattleSceneResources(options: {
     playerArmdozerIds,
     resourceArmdozerIds,
   });
-  const shouldLoading = (t: ResourceType) =>
-    t.type === "DynamicArmdozer" &&
-    additionalArmdozerIds.includes(t.armdozerId);
-  const additionalLoading = loadResources({
-    resourceRoot: resources.rootPath,
-    preLoadImages: [],
-    gltfConfigs: GLTF_CONFIGS.filter(shouldLoading),
-    textureConfigs: TEXTURE_CONFIGS.filter(shouldLoading),
-    cubeTextureConfigs: CUBE_TEXTURE_CONFIGS.filter(shouldLoading),
-    canvasImageConfigs: CANVAS_IMAGE_CONFIGS.filter(shouldLoading),
-    soundConfigs: SOUND_CONFIGS.filter(shouldLoading),
+  const additonalResources = await loadAdditionalResources({
+    resources,
+    additionalArmdozerIds,
   });
-  const loaded = await additionalLoading.resources;
-  return mergeResources({ resources, loaded });
+  return mergeResources({ resources, loaded: additonalResources });
 }
